requirements.txt
Django==5.0.14
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
Pillow==10.2.0
python-dotenv==1.0.1
